cleaned up comments, headers
