+++ Sources/ReleaseTools/ApiKeyOption.swift
@@ -1,0 +1,12 @@
+// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
+//  Created by Sam Deane on 07/04/2026.
+//  Copyright © 2026 Elegant Chaos Limited. All rights reserved.
+// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
+
+import ArgumentParser
+
+/// Common command-line option for selecting an App Store Connect API key.
+struct ApiKeyOption: ParsableArguments {
+  @Option(name: .customLong("api-key"), help: "The App Store Connect api key ID we're using.")
+  var key: String?
+}
