+++ Sources/ReleaseTools/PlatformOption.swift
@@ -1,0 +1,12 @@
+// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
+//  Created by Sam Deane on 07/04/2026.
+//  Copyright © 2026 Elegant Chaos Limited. All rights reserved.
+// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
+
+import ArgumentParser
+
+/// Common command-line option for selecting a target platform.
+struct PlatformOption: ParsableArguments {
+  @Option(help: "The platform to build for. Should be one of: macOS, iOS, tvOS, watchOS.")
+  var platform: String?
+}
