+++ Sources/ReleaseTools/RTConfigReader.swift
@@ -1,14 +1,16 @@
 // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
-//  Created by Sam Deane on 20/03/26.
-//  All code (c) 2020 - present day, Elegant Chaos Limited.
+//  Created by Sam Deane on 20/03/2026.
+//  Copyright © 2026 Elegant Chaos Limited. All rights reserved.
 // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
 
 import Configuration
 import Foundation
 
+/// Loads layered configuration snapshots from the paths resolved by `RTConfigPaths`.
 struct RTConfigReader {
   let config: ConfigReader
 
+  /// Creates a reader using the candidate config files for the active scheme and platform.
   init(paths: RTConfigPaths, scheme: String?, platform: String?) async throws {
     var providers: [any ConfigProvider] = []
     for url in paths.candidateURLs(scheme: scheme, platform: platform) {
@@ -29,10 +31,12 @@ struct RTConfigReader {
     self.config = ConfigReader(providers: providers)
   }
 
+  /// The default scheme from layered configuration.
   var defaultScheme: String? {
     config.string(forKey: "defaults.scheme")
   }
 
+  /// Effective release settings from layered configuration.
   var settings: BasicSettings {
     BasicSettings(
       keychain: config.string(forKey: "settings.keychain"),
