+++ Sources/ReleaseTools/Runners/DittoRunner.swift
@@ -1,18 +1,22 @@
 // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
-//  Created by Sam Deane on 24/02/20.
-//  All code (c) 2020 - present day, Elegant Chaos Limited.
+//  Created by Sam Deane on 24/02/2020.
+//  Copyright © 2020 Elegant Chaos Limited. All rights reserved.
 // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
 
 import Foundation
 import Runner
 
-class DittoRunner: Runner {
+/// `ditto` runner used for archive compression.
+final class DittoRunner: Runner {
   let engine: ReleaseEngine
+
+  /// Creates a runner bound to one release engine instance.
   init(engine: ReleaseEngine) {
     self.engine = engine
     super.init(command: "ditto")
   }
 
+  /// Runs a `ditto` command using the engine's output policy.
   func run(_ arguments: [String]) -> Session {
     if engine.showCommands {
       engine.log("ditto " + arguments.joined(separator: " "))
@@ -22,6 +26,7 @@ class DittoRunner: Runner {
     return run(arguments, stdoutMode: mode, stderrMode: mode)
   }
 
+  /// Compresses a bundle into a zip archive using `ditto`.
   func zip(_ url: URL, as zipURL: URL) -> Session {
     engine.log("Compressing \(url.lastPathComponent) to \(zipURL.path).")
     return run(["-c", "-k", "--sequesterRsrc", "--keepParent", url.path, zipURL.path])
