+++ Sources/ReleaseTools/SchemeOption.swift
@@ -1,0 +1,12 @@
+// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
+//  Created by Sam Deane on 07/04/2026.
+//  Copyright © 2026 Elegant Chaos Limited. All rights reserved.
+// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
+
+import ArgumentParser
+
+/// Common command-line option for selecting an Xcode scheme.
+struct SchemeOption: ParsableArguments {
+  @Option(help: "The scheme we're building.")
+  var scheme: String?
+}
